move tests to opts/triton/test_ssd.py
