+++ tests/ops/test_selective_scan.py
@@ -1,7 +1,6 @@
 # Copyright (C) 2023, Tri Dao.
 
-from typing import Optional
-from copy import deepcopy
+import math
 
 import torch
 import torch.nn.functional as F
@@ -11,9 +10,6 @@ from einops import rearrange
 
 from mamba_ssm.ops.selective_scan_interface import selective_scan_fn, selective_scan_ref
 from mamba_ssm.ops.selective_scan_interface import mamba_inner_fn, mamba_inner_ref
-from mamba_ssm.ops.triton.ssd_combined import mamba_chunk_scan_combined, mamba_split_conv1d_scan_combined
-from mamba_ssm.modules.ssd_minimal import ssd_minimal_discrete
-from mamba_ssm.modules.mamba2 import Mamba2
 
 
 # @pytest.mark.parametrize('wtype', [torch.float32, torch.complex64])
@@ -249,347 +245,3 @@ def test_mamba_inner_fn(is_variable_B, is_variable_C, seqlen, itype, wtype):
     #                       atol=atolw if not is_variable_C else atol)
     # assert torch.allclose(D.grad, D_ref.grad, rtol=rtolw, atol=atolw)
     # assert torch.allclose(delta_bias.grad, delta_bias_ref.grad, rtol=rtolw, atol=atolw)
-
-
-
-def get_seq_idx_and_cu_seqlens(
-    max_splits: int, seqlen: int, device: torch.device
-)->tuple[torch.Tensor, torch.Tensor]:
-    nsplits = torch.randint(1, max_splits + 1, (1,)).item()
-    eos_pos = torch.randperm(seqlen - 1)[:nsplits].sort().values
-    cu_seqlens = (
-        torch.cat([torch.tensor([-1]), eos_pos, torch.tensor([seqlen - 1])]) + 1
-    )
-    seqlens = torch.diff(cu_seqlens).tolist()
-    assert sum(seqlens) == seqlen
-    assert all(s > 0 for s in seqlens)
-    seq_idx = torch.stack(
-        [
-            torch.cat(
-                [
-                    torch.full((s,), i, dtype=torch.int32, device=device)
-                    for i, s in enumerate(seqlens)
-                ],
-                dim=0,
-            )
-        ],
-        dim=0,
-    )
-    return seq_idx, cu_seqlens
-
-
-class TestMambaChunkScanCombined:
-    seqlen = 256
-    chunk_size = 32
-    dim = 128
-    headdim = 32
-    nheads = dim // headdim
-    ngroups = 1
-    dstate = 8
-    dtype = torch.float32
-    device = "cuda"
-    max_splits = 4
-
-    def _get_xdtABC(self, requires_grad: bool = False, batch_size: int = 1):
-        x = torch.randn(
-            batch_size,
-            self.seqlen,
-            self.nheads,
-            self.headdim,
-            dtype=self.dtype,
-            device=self.device,
-            requires_grad=requires_grad,
-        )
-        dt = F.softplus(
-            torch.randn(
-                batch_size,
-                self.seqlen,
-                self.nheads,
-                dtype=self.dtype,
-                device=self.device,
-            )
-            - 4
-        )
-        A = -torch.exp(
-            torch.rand(
-                self.nheads,
-                dtype=self.dtype,
-                device=self.device,
-            )
-        )
-        if requires_grad:
-            # Set dt and A as requires_grad, and not the tensors they're built from, so that they
-            # are leaf tensors which accumulate gradients.
-            dt.requires_grad_()
-            A.requires_grad_()
-        B = torch.randn(
-            batch_size,
-            self.seqlen,
-            self.ngroups,
-            self.dstate,
-            dtype=self.dtype,
-            device=self.device,
-            requires_grad=requires_grad,
-        )
-        C = torch.randn(
-            batch_size,
-            self.seqlen,
-            self.ngroups,
-            self.dstate,
-            dtype=self.dtype,
-            device=self.device,
-            requires_grad=requires_grad,
-        )
-        return x, dt, A, B, C
-
-    def test_fwd(self) -> None:
-        """
-        Test the triton mamba_chunk_scan_combined against the pure torch implementation
-        ssd_minimal_discrete.
-        """
-        torch.manual_seed(42)
-        x, dt, A, B, C = self._get_xdtABC()
-        y = mamba_chunk_scan_combined(x, dt, A, B, C, self.chunk_size, D=None)
-        y_min, _ = ssd_minimal_discrete(
-            x * dt.unsqueeze(-1), A * dt, B, C, self.chunk_size
-        )
-        # These tolerances seem high, but the test fails for rtol = atol = 1e-3. Surprising?
-        rtol = atol = 1e-2
-        assert torch.allclose(y, y_min, rtol=rtol, atol=atol)
-
-    def test_bwd(self) -> None:
-        """
-        Test the triton mamba_chunk_scan_combined against the pure torch implementation
-        ssd_minimal_discrete with a backwards pass.
-        """
-        torch.manual_seed(42)
-        x, dt, A, B, C = self._get_xdtABC(requires_grad=True)
-
-        x_c = x.detach().clone().requires_grad_()
-        dt_c = dt.detach().clone().requires_grad_()
-        A_c = A.detach().clone().requires_grad_()
-        B_c = B.detach().clone().requires_grad_()
-        C_c = C.detach().clone().requires_grad_()
-
-        y = mamba_chunk_scan_combined(x, dt, A, B, C, self.chunk_size, D=None)
-        y_c, _ = ssd_minimal_discrete(
-            x_c * dt_c.unsqueeze(-1), A_c * dt_c, B_c, C_c, self.chunk_size
-        )
-
-        y.sum().backward()
-        y_c.sum().backward()
-
-        # Test only passes with large tolerances. rtol=atol=1e-2 fails. The dt and C grads have
-        # largest discrepancies. Surprising?
-        rtol = atol = 1e-1
-        with torch.no_grad():
-            assert torch.allclose(x.grad, x_c.grad, rtol=rtol, atol=atol)
-            assert torch.allclose(dt.grad, dt_c.grad, rtol=rtol, atol=atol)
-            assert torch.allclose(A.grad, A_c.grad, rtol=rtol, atol=atol)
-            assert torch.allclose(B.grad, B_c.grad, rtol=rtol, atol=atol)
-            assert torch.allclose(C.grad, C_c.grad, rtol=rtol, atol=atol)
-
-    def test_seq_idx_fwd(self) -> None:
-        """
-        Similar to causal-conv1d's test_causal_conv1d_varlen.
-        """
-        torch.manual_seed(42)
-        x, dt, A, B, C = self._get_xdtABC()
-        seq_idx, cu_seqlens = get_seq_idx_and_cu_seqlens(
-            self.max_splits, self.seqlen, self.device
-        )
-
-        y = mamba_chunk_scan_combined(
-            x, dt, A, B, C, self.chunk_size, D=None, seq_idx=seq_idx
-        )
-        atol = rtol = 1e-3
-        start_idxs = cu_seqlens[:-1]
-        stop_idxs = cu_seqlens[1:]
-        for start_idx, stop_idx in zip(start_idxs, stop_idxs):
-            x_chunk = x[:, start_idx:stop_idx]
-            dt_chunk = dt[:, start_idx:stop_idx]
-            B_chunk = B[:, start_idx:stop_idx]
-            C_chunk = C[:, start_idx:stop_idx]
-            y_chunk = mamba_chunk_scan_combined(
-                x_chunk, dt_chunk, A, B_chunk, C_chunk, self.chunk_size, D=None
-            )
-            y_chunk_expected = y[:, start_idx:stop_idx]
-            assert torch.allclose(y_chunk, y_chunk_expected, rtol=rtol, atol=atol)
-
-    def test_seq_idx_bwd(self) -> None:
-        # HACK: failed on ~1% of elements with seed 42, but passes with 43.
-        torch.manual_seed(43)
-        x, dt, A, B, C = self._get_xdtABC(requires_grad=True)
-
-        seq_idx, cu_seqlens = get_seq_idx_and_cu_seqlens(
-            self.max_splits, self.seqlen, self.device
-        )
-        y = mamba_chunk_scan_combined(
-            x, dt, A, B, C, self.chunk_size, D=None, seq_idx=seq_idx
-        )
-        y.sum().backward()
-
-        atol = rtol = 1e-2
-        start_idxs = cu_seqlens[:-1]
-        stop_idxs = cu_seqlens[1:]
-        A_grads = torch.zeros_like(A)
-        for start_idx, stop_idx in zip(start_idxs, stop_idxs):
-            x_chunk = x[:, start_idx:stop_idx].detach().clone().requires_grad_()
-            dt_chunk = dt[:, start_idx:stop_idx].detach().clone().requires_grad_()
-            B_chunk = B[:, start_idx:stop_idx].detach().clone().requires_grad_()
-            C_chunk = C[:, start_idx:stop_idx].detach().clone().requires_grad_()
-            A_copy = A.detach().clone().requires_grad_()
-            y_chunk = mamba_chunk_scan_combined(
-                x_chunk, dt_chunk, A_copy, B_chunk, C_chunk, self.chunk_size, D=None
-            )
-            y_chunk.sum().backward()
-
-            # Need to extract the grad first, then slice
-            x_chunk_expected_grad = x.grad[:, start_idx:stop_idx]
-            assert torch.allclose(
-                x_chunk.grad, x_chunk_expected_grad, rtol=rtol, atol=atol
-            )
-            dt_chunk_expected_grad = dt.grad[:, start_idx:stop_idx]
-            assert torch.allclose(
-                dt_chunk.grad, dt_chunk_expected_grad, rtol=rtol, atol=atol
-            )
-            B_chunk_expected_grad = B.grad[:, start_idx:stop_idx]
-            assert torch.allclose(
-                B_chunk.grad, B_chunk_expected_grad, rtol=rtol, atol=atol
-            )
-            C_chunk_expected_grad = C.grad[:, start_idx:stop_idx]
-            assert torch.allclose(
-                C_chunk.grad, C_chunk_expected_grad, rtol=rtol, atol=atol
-            )
-            A_grads += A_copy.grad
-        assert torch.allclose(A_grads, A.grad, rtol=rtol, atol=atol)
-
-
-class TestMambaSplitConv1dScanCombined:
-    seqlen = 256
-    chunk_size = 32
-    d_model = 128
-    headdim = 32
-    d_state = 8
-    expand = 2
-    ngroups = 1
-    d_inner = expand * d_model
-    nheads = d_inner // headdim
-    dtype = torch.float32
-    device = "cuda"
-    max_splits = 4
-
-    def _get_model(self, seed: Optional[int] = None) -> Mamba2:
-        if seed is not None:
-            torch.manual_seed(seed)
-        mamba2 = Mamba2(
-            d_model=self.d_model,
-            device=self.device,
-            chunk_size=self.chunk_size,
-            headdim=self.headdim,
-            d_state=self.d_state,
-            ngroups=self.ngroups,
-            expand=self.expand,
-        )
-        return mamba2
-
-    def _get_kwargs_from_model(self, model: Mamba2) -> dict:
-        kwargs = {
-            "conv1d_weight": rearrange(model.conv1d.weight, "d 1 w -> d w"),
-            "conv1d_bias": model.conv1d.bias,
-            "dt_bias": model.dt_bias,
-            "A": -torch.exp(model.A_log.float()),
-            "D": rearrange(model.D, "(h p) -> h p", p=model.headdim)
-            if model.D_has_hdim
-            else model.D,
-            "chunk_size": model.chunk_size,
-            "activation": model.activation,
-            "rmsnorm_weight": model.norm.weight if model.rmsnorm else None,
-            "rmsnorm_eps": model.norm.eps if model.rmsnorm else 1e-6,
-            "outproj_weight": model.out_proj.weight,
-            "outproj_bias": model.out_proj.bias,
-            "headdim": None if model.D_has_hdim else model.headdim,
-            "ngroups": model.ngroups,
-            "norm_before_gate": model.norm_before_gate,
-        }
-        return kwargs
-
-    def _get_zxbcdt(
-        self,
-        requires_grad: bool = False,
-        batch_size: int = 1,
-        seed: Optional[int] = None,
-    ) -> torch.Tensor:
-        if seed is not None:
-            torch.manual_seed(seed)
-        d_in_proj = 2 * self.d_inner + 2 * self.ngroups * self.d_state + self.nheads
-        zxbcdt = torch.randn(
-            batch_size,
-            self.seqlen,
-            d_in_proj,
-            dtype=self.dtype,
-            device=self.device,
-            requires_grad=requires_grad,
-        )
-        return zxbcdt
-
-    def test_fwd_seq_idx(self) -> None:
-        seed = 42
-        model = self._get_model(seed=seed)
-        zxbcdt = self._get_zxbcdt(seed=seed)
-        seq_idx, cu_seqlens = get_seq_idx_and_cu_seqlens(
-            self.max_splits, self.seqlen, self.device
-        )
-
-        kwargs = self._get_kwargs_from_model(model)
-        y = mamba_split_conv1d_scan_combined(zxbcdt, seq_idx=seq_idx, **kwargs)
-
-        atol = rtol = 1e-3
-        start_idxs = cu_seqlens[:-1]
-        stop_idxs = cu_seqlens[1:]
-        for start_idx, stop_idx in zip(start_idxs, stop_idxs):
-            zxbcdt_chunk = zxbcdt[:, start_idx:stop_idx]
-            y_chunk = mamba_split_conv1d_scan_combined(zxbcdt_chunk, **kwargs)
-            y_chunk_expected = y[:, start_idx:stop_idx]
-            assert torch.allclose(y_chunk, y_chunk_expected, rtol=rtol, atol=atol)
-
-    def test_bwd_seq_idx(self) -> None:
-        seed = 42
-        model = self._get_model(seed=seed)
-        model_c = deepcopy(model)
-        zxbcdt = self._get_zxbcdt(seed=seed, requires_grad=True)
-        seq_idx, cu_seqlens = get_seq_idx_and_cu_seqlens(
-            self.max_splits, self.seqlen, self.device
-        )
-
-        kwargs = self._get_kwargs_from_model(model)
-        y = mamba_split_conv1d_scan_combined(zxbcdt, seq_idx=seq_idx, **kwargs)
-        y.sum().backward()
-
-        atol = rtol = 1e-3
-        start_idxs = cu_seqlens[:-1]
-        stop_idxs = cu_seqlens[1:]
-        for start_idx, stop_idx in zip(start_idxs, stop_idxs):
-            kwargs_c = self._get_kwargs_from_model(model_c)
-            # Create chunk with requires_grad=False, then slice, then requires_grad_, so that it's a
-            # leaf tensor which accumulates grads.
-            zxbcdt_chunk = self._get_zxbcdt(seed=seed)[
-                :, start_idx:stop_idx
-            ].requires_grad_()
-            y_chunk = mamba_split_conv1d_scan_combined(zxbcdt_chunk, **kwargs_c)
-            y_chunk.sum().backward()
-            zxbcdt_chunk_grad_expected = zxbcdt.grad[:, start_idx:stop_idx]
-            assert torch.allclose(
-                zxbcdt_chunk.grad,
-                zxbcdt_chunk_grad_expected,
-                rtol=rtol,
-                atol=atol,
-            )
-
-        for p1, (n, p2) in zip(model_c.parameters(), model.named_parameters()):
-            if p2.grad is None:
-                assert p1.grad is None, f"{n=}"
-            else:
-                assert torch.allclose(
-                    p1.grad, p2.grad, rtol=rtol, atol=atol
-                ), f"Failed on {n=}"
